Move mailboxes code to separate app
