+++ conf/settings.py
@@ -47,6 +47,7 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "mail",
+    "mailboxes",
     "health_check",
     "health_check.db",
     "health_check.cache",
