+++ mailboxes/enums.py
@@ -1,0 +1,7 @@
+from django.db.models import TextChoices
+
+
+class MailReadStatuses(TextChoices):
+    READ = "READ"
+    UNREAD = "UNREAD"
+    UNPROCESSABLE = "UNPROCESSABLE"
